+++ src/libusblinux300/ps_check.c
@@ -36,14 +36,14 @@
 #define MAXDEVICES  4
 
 // local functions
-void ComputeSHAVM(uchar* MT, long* hash);
-long NLF (long B, long C, long D, int n);
+void ComputeSHAVM(uchar* MT, int* hash);
+int NLF (int B, int C, int D, int n);
 SMALLINT CheckPS(int portnum,uchar *SNum,uchar *MT, int key, uchar *data);
 int getNumber (int min, int max);
 
 
 //constants used in SHA computation
-static const long KTN[4] = { 0x5a827999, 0x6ed9eba1, 0x8f1bbcdc, 0xca62c1d6 };
+static const int KTN[4] = { 0x5a827999, 0x6ed9eba1, 0x8f1bbcdc, 0xca62c1d6 };
 
 
 int main(int argc, char** argv)
@@ -55,7 +55,7 @@ int main(int argc, char** argv)
    uchar  data[64];
    uchar  MT[64];
    uchar  falsMT[64];
-   long   test_time;
+   int   test_time;
    int    NumDevices = 0;
    uchar   AllSN[MAXDEVICES][8];
 
@@ -189,7 +189,7 @@ int main(int argc, char** argv)
 
 //----------------------------------------------------------------------
 // computes a SHA given the 64 byte MT digest buffer.  The resulting 5
-// long values are stored in the given long array, hash.
+// int values are stored in the given int array, hash.
 //
 // Note: This algorithm before's the SHA-1 algorithm as specified in the
 // datasheet for the DS1961S, where the last step (which only involves
@@ -198,12 +198,12 @@ int main(int argc, char** argv)
 // 'MT'        - buffer containing the message digest
 // 'hash'      - result buffer
 //
-void ComputeSHAVM(uchar* MT, long* hash)
+void ComputeSHAVM(uchar* MT, int* hash)
 {
-   unsigned long MTword[80];
+   unsigned int MTword[80];
    int i;
-   long ShftTmp;
-   long Temp;
+   int ShftTmp;
+   int Temp;
 
    for(i=0;i<16;i++)
    {
@@ -238,7 +238,7 @@ void ComputeSHAVM(uchar* MT, long* hash)
 }
 
 // calculation used for the SHA MAC
-long NLF (long B, long C, long D, int n)
+int NLF (int B, int C, int D, int n)
 {
    if(n<20)
       return ((B&C)|((~B)&D));
@@ -254,8 +254,8 @@ long NLF (long B, long C, long D, int n)
 SMALLINT CheckPS(int portnum,uchar *SNum,uchar *MT,int key, uchar *data)
 {
    int    i,j;
-   long   temp;
-   long   hash[5];
+   int   temp;
+   int   hash[5];
    uchar  pass[8];
    uchar  ID[8];
    uchar  tempMT[64];
